requirement.txt
streamlit
scikit-learn
pandas
numpy
Pillow

